Reorganize code to minimize header bloat
