+++ Test/gsapTests/Loading/GaussianLoadEstimatorTests.cpp
@@ -1,6 +1,7 @@
 // Copyright (c) 2018 United States Government as represented by the
 // Administrator of the National Aeronautics and Space Administration.
 // All Rights Reserved.
+#include "ConfigMap.h"
 #include "Loading/GaussianLoadEstimator.h"
 #include "Loading/LoadEstimatorTests.h"
