+++ inc/Loading/MovingAverageLoadEstimator.h
@@ -3,12 +3,11 @@
 // All Rights Reserved.
 #ifndef PCOE_MOVINAVERAGELOADESTIMATOR_H
 #define PCOE_MOVINAVERAGELOADESTIMATOR_H
-
-#include "ConfigMap.h"
-#include "Contracts.h"
 #include "Loading/LoadEstimator.h"
 
 namespace PCOE {
+    class ConfigMap;
+
     /**
      * Produces a constant load estimate that is the average of several previous
      * loads.
@@ -19,9 +18,8 @@ namespace PCOE {
      **/
     class MovingAverageLoadEstimator final : public LoadEstimator {
     public:
-        static const std::string
-            WINDOW_SIZE_KEY; // Key for window size (number of steps in average window)
-        static const size_t DEFAULT_WINDOW_SIZE;
+        // The window is the number of samples considered in the moving average.
+        static const std::string WINDOW_SIZE_KEY;
 
         /**
          * Constructs a new {@code MovingAverageLoadEstimator} instance.
@@ -38,25 +36,7 @@ namespace PCOE {
          * @param config The configuration used to initialize the load
          *               estimator.
          **/
-        MovingAverageLoadEstimator(const ConfigMap& config) {
-            const std::string LOADING_KEY = "LoadEstimator.Loading";
-            const std::string WINDOW_KEY = "LoadEstimator.Window";
-            requireKeys(config, {LOADING_KEY});
-
-            std::size_t window = 10;
-            if (config.hasKey(WINDOW_KEY)) {
-                window = config.getSize(WINDOW_KEY);
-            }
-
-            currentEstimate = config.getDoubleVector(LOADING_KEY);
-
-            std::vector<double> partialEstimate(currentEstimate.size());
-            for (std::size_t i = 0; i < currentEstimate.size(); ++i) {
-                partialEstimate[i] = currentEstimate[i] / window;
-            }
-
-            pastEstimates.resize(window, partialEstimate);
-        }
+        MovingAverageLoadEstimator(const ConfigMap& config);
 
         /**
          * Returns a value indicating that the load estimator supports
@@ -72,16 +52,7 @@ namespace PCOE {
          *
          * @param load The new load to add.
          **/
-        void addLoad(const LoadEstimate& load) override {
-            Expect(load.size() == currentEstimate.size(), "Size mismatch");
-            for (std::size_t i = 0; i < load.size(); ++i) {
-                double newEst = load[i] / pastEstimates.size();
-                currentEstimate[i] -= pastEstimates[pos][i];
-                currentEstimate[i] += newEst;
-                pastEstimates[pos][i] = newEst;
-            }
-            pos = (pos + 1) % pastEstimates.size();
-        }
+        void addLoad(const LoadEstimate& load) override;
 
         /**
          * Gets the current load estimate.
@@ -89,11 +60,7 @@ namespace PCOE {
          * @param t      Not used.
          * @return       The current estimated load.
          **/
-        LoadEstimate estimateLoad(const double t) override {
-            static_cast<void>(t);
-
-            return currentEstimate;
-        }
+        LoadEstimate estimateLoad(const double t) override;
 
     protected:
         std::size_t pos = 0;
