+++ inc/Messages/MessageWatcher.h
@@ -7,6 +7,7 @@
 
 #include "Contracts.h"
 #include "Messages/MessageBus.h"
+#include "Messages/VectorMessage.h"
 
 namespace PCOE {
     /**
